backend/requirements.txt
# Core
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# AI Providers
anthropic>=0.40.0
openai>=1.50.0

# Database
sqlalchemy>=2.0.0
aiosqlite>=0.19.0

# Vector Store (RAG)
chromadb>=0.4.0

# WebSocket & Streaming
websockets>=12.0
sse-starlette>=2.0.0

# File handling
python-multipart>=0.0.9
aiofiles>=23.0
pypdf2>=3.0.0
python-docx>=1.0.0

# HTTP client
httpx>=0.27.0

# Utilities
rich>=13.0.0
